Remove custom console system - use standard logging
Removes the custom console module (9 files) and replaces it with
standard Python logging and Rich library imports:

- Replace get_logger() with logging.getLogger()
- Replace print_* functions with logger.info/error/warning calls
- Move RichProgressWrapper to progress module
- Update util/logging_config.py to use Rich directly
- Update example files to import from Rich directly

Simplifies codebase by removing 500+ lines of custom wrapper code
and using battle-tested standard libraries instead.

Fixes #107
