+++ socialmapper/geocoding/providers.py
@@ -4,6 +4,7 @@
 This module contains implementations of various geocoding providers.
 """
 
+import logging
 import time
 from abc import ABC, abstractmethod
 from typing import Any
@@ -12,11 +13,10 @@ import requests
 from requests.adapters import HTTPAdapter
 from urllib3.util.retry import Retry
 
-from ..console import get_logger
 from ..neighbors import get_neighbor_manager
 from .models import AddressInput, AddressProvider, AddressQuality, GeocodingConfig, GeocodingResult
 
-logger = get_logger(__name__)
+logger = logging.getLogger(__name__)
 
 
 class GeocodingProvider(ABC):
