+++ socialmapper/io/manager.py
@@ -1,14 +1,13 @@
 """Central I/O manager for tracking and organizing all file operations."""
 
 import json
+import logging
 from dataclasses import dataclass, field
 from datetime import datetime
 from pathlib import Path
 from typing import Any
 
-from ..console import get_logger
-
-logger = get_logger(__name__)
+logger = logging.getLogger(__name__)
 
 
 @dataclass
