+++ socialmapper/isochrone/__init__.py
@@ -13,6 +13,7 @@ Key Features:
 """
 
 import json
+import logging
 import os
 import time
 import warnings
@@ -25,8 +26,6 @@ import networkx as nx
 import osmnx as ox
 import pandas as pd
 from shapely.geometry import Point
-
-from socialmapper.console import get_logger
 
 # Import the new progress bar utility
 from socialmapper.progress import get_progress_bar
@@ -45,7 +44,7 @@ from .concurrent import ConcurrentIsochroneProcessor, ProcessingStats, process_i
 from .travel_modes import TravelMode, get_travel_mode_config
 
 # Setup logging
-logger = get_logger(__name__)
+logger = logging.getLogger(__name__)
 
 # Suppress FutureWarning
 warnings.filterwarnings("ignore", category=FutureWarning)
