+++ socialmapper/pipeline/extraction.py
@@ -160,9 +160,9 @@ def parse_custom_coordinates(
                     pois.append(poi)
                 else:
                     # Log warning but don't fail - some POIs might be malformed
-                    from ..console import print_warning
-
-                    print_warning(f"Skipping item missing required coordinates: {item}")
+                    import logging
+                    logger = logging.getLogger(__name__)
+                    logger.warning(f"Skipping item missing required coordinates: {item}")
         elif isinstance(data, dict) and "pois" in data:
             pois = data["pois"]
