+++ socialmapper/util/logging_config.py
@@ -6,7 +6,7 @@ This module sets up Rich-enhanced logging for the entire package.
 import logging
 import os
 
-from ..console import setup_rich_logging
+from rich.logging import RichHandler
 
 
 def configure_logging(level=None):
@@ -20,12 +20,23 @@ def configure_logging(level=None):
         env_level = os.environ.get("SOCIALMAPPER_LOG_LEVEL", "CRITICAL").upper()
         level = env_level
 
-    # Use Rich logging setup
-    setup_rich_logging(level=level, show_time=False, show_path=False)
+    # Convert string level to int if needed
+    if isinstance(level, str):
+        level_int = getattr(logging, level.upper(), logging.CRITICAL)
+    else:
+        level_int = level
+
+    # Configure root logger with Rich handler
+    rich_handler = RichHandler(show_time=False, show_path=False, markup=True, rich_tracebacks=True)
+    logging.basicConfig(
+        level=level_int,
+        format="%(message)s",
+        handlers=[rich_handler]
+    )
 
     # Set socialmapper logger level
     logger = logging.getLogger("socialmapper")
-    logger.setLevel(getattr(logging, level, logging.CRITICAL))
+    logger.setLevel(level_int)
 
     # Also set urllib3 to warning or higher to reduce noise
     logging.getLogger("urllib3").setLevel(logging.WARNING)
